backend/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.27
alembic>=1.13.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.27.0
redis>=5.0.1
scikit-learn>=1.6.0
numpy>=1.26.4
pandas>=2.2.0
apscheduler>=3.10.4
aiosqlite>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.23.4
